refactor out ethereum logic into own crates
