+++ crates/core-ethereum/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "zeth-core-ethereum"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies.zeth-core]
+workspace = true
+
+[dependencies]
+anyhow.workspace = true
+
+alloy-consensus.workspace = true
+
+reth-chainspec.workspace = true
+reth-consensus.workspace = true
+reth-ethereum-consensus.workspace = true
+reth-evm.workspace = true
+reth-evm-ethereum.workspace = true
+reth-primitives.workspace = true
+reth-revm.workspace = true
+reth-storage-errors.workspace = true
