+++ guests/reth/src/main.rs
@@ -16,7 +16,8 @@ use reth_chainspec::MAINNET;
 // use c_kzg::KzgSettings;
 use risc0_zkvm::guest::env;
 use risc0_zkvm::guest::env::stdin;
-use zeth_core::stateless::client::{RethStatelessClient, StatelessClient};
+use zeth_core::stateless::client::StatelessClient;
+use zeth_core_ethereum::RethStatelessClient;
 // todo: use this instead of the alloy KzgEnv to save cycles
 // lazy_static::lazy_static! {
 //     /// KZG Ceremony data
